services/it-support-crew/requirements.txt
flask
crewai
python-dotenv
google-generativeai
paramiko
pyyaml
requests
langchain
langtrace-python-sdk
pydantic